requirements.txt
openai>=1.0.0
anthropic>=0.20.0
pyyaml>=6.0
